+++ lib/plausible/exports.ex
@@ -130,26 +130,18 @@ defmodule Plausible.Exports do
 
   @spec local_export_file(pos_integer) :: Path.t()
   defp local_export_file(site_id) do
-    persistent_cache_dir = Application.get_env(:plausible, :persistent_cache_dir)
-
-    Path.join([
-      persistent_cache_dir || System.tmp_dir!(),
-      "plausible-exports",
-      Integer.to_string(site_id)
-    ])
+    data_dir = Application.get_env(:plausible, :data_dir)
+    Path.join([data_dir || System.tmp_dir!(), "plausible-exports", Integer.to_string(site_id)])
   end
 
-  @doc "Gets S3 export for a site"
-  @spec get_s3_export(pos_integer) :: export | nil
-  def get_s3_export(site_id) do
+  @doc "Gets S3 export for a site. Raises if object storage is unavailable."
+  @spec get_s3_export!(pos_integer, non_neg_integer) :: export | nil
+  def get_s3_export!(site_id, retries \\ 0) do
     path = s3_export_key(site_id)
     bucket = Plausible.S3.exports_bucket()
     head_object_op = ExAws.S3.head_object(bucket, path)
 
-    case ExAws.request(head_object_op) do
-      {:error, {:http_error, 404, _response}} ->
-        nil
-
+    case ExAws.request(head_object_op, retries: retries) do
       {:ok, %{status_code: 200, headers: headers}} ->
         "attachment; filename=" <> filename = :proplists.get_value("content-disposition", headers)
         name = String.trim(filename, "\"")
@@ -169,13 +161,19 @@ defmodule Plausible.Exports do
           expires_at: expires_at,
           size: String.to_integer(size)
         }
+
+      {:error, {:http_error, 404, _response}} ->
+        nil
+
+      {:error, %Mint.TransportError{} = e} ->
+        raise e
     end
   end
 
-  @doc "Deletes S3 export for a site"
-  @spec delete_s3_export(pos_integer) :: :ok
-  def delete_s3_export(site_id) do
-    if export = get_s3_export(site_id) do
+  @doc "Deletes S3 export for a site. Raises if object storage is unavailable."
+  @spec delete_s3_export!(pos_integer) :: :ok
+  def delete_s3_export!(site_id) do
+    if export = get_s3_export!(site_id) do
       exports_bucket = Plausible.S3.exports_bucket()
       delete_op = ExAws.S3.delete_object(exports_bucket, export.path)
       ExAws.request!(delete_op)
@@ -187,7 +185,7 @@ defmodule Plausible.Exports do
   defp s3_export_key(site_id), do: Integer.to_string(site_id)
 
   @doc "Returns the date range for the site's events data in site's timezone or `nil` if there is no data"
-  @spec date_range(non_neg_integer, String.t()) :: Date.Range.t() | nil
+  @spec date_range(pos_integer, String.t()) :: Date.Range.t() | nil
   def date_range(site_id, timezone) do
     [%Date{} = start_date, %Date{} = end_date] =
       Plausible.ClickhouseRepo.one(
@@ -245,24 +243,27 @@ defmodule Plausible.Exports do
   end
 
   on_ee do
-    defp sampled(table, date_range) do
-      from(table)
-      |> Plausible.Stats.Sampling.add_query_hint()
-      |> limit_date_range(date_range)
+    defp sampled(table) do
+      Plausible.Stats.Sampling.add_query_hint(from(table))
     end
   else
-    defp sampled(table, date_range) do
-      limit_date_range(table, date_range)
+    defp sampled(table) do
+      table
     end
   end
 
-  defp limit_date_range(query, nil), do: query
+  defp export_filter(site_id, date_range) do
+    filter = dynamic([t], t.site_id == ^site_id)
 
-  defp limit_date_range(query, date_range) do
-    from t in query,
-      where:
-        selected_as(:date) >= ^date_range.first and
+    if date_range do
+      dynamic(
+        ^filter and
+          selected_as(:date) >= ^date_range.first and
           selected_as(:date) <= ^date_range.last
+      )
+    else
+      filter
+    end
   end
 
   defmacrop date(timestamp, timezone) do
@@ -277,7 +278,11 @@ defmodule Plausible.Exports do
   defmacrop visit_duration(t) do
     quote do
       selected_as(
-        fragment("greatest(sum(?*?),0)", unquote(t).sign, unquote(t).duration),
+        fragment(
+          "toUInt64(round(greatest(sum(?*?),0)*any(_sample_factor)))",
+          unquote(t).sign,
+          unquote(t).duration
+        ),
         :visit_duration
       )
     end
@@ -294,14 +299,21 @@ defmodule Plausible.Exports do
 
   defmacrop visits(t) do
     quote do
-      selected_as(sum(unquote(t).sign), :visits)
+      selected_as(
+        fragment("toUInt64(round(greatest(sum(?),0)*any(_sample_factor)))", unquote(t).sign),
+        :visits
+      )
     end
   end
 
   defmacrop bounces(t) do
     quote do
       selected_as(
-        fragment("greatest(sum(?*?),0)", unquote(t).sign, unquote(t).is_bounce),
+        fragment(
+          "toUInt32(round(greatest(sum(?*?),0)*any(_sample_factor)))",
+          unquote(t).sign,
+          unquote(t).is_bounce
+        ),
         :bounces
       )
     end
@@ -310,7 +322,11 @@ defmodule Plausible.Exports do
   defmacrop pageviews(t) do
     quote do
       selected_as(
-        fragment("greatest(sum(?*?),0)", unquote(t).sign, unquote(t).pageviews),
+        fragment(
+          "toUInt64(round(greatest(sum(?*?),0)*any(_sample_factor)))",
+          unquote(t).sign,
+          unquote(t).pageviews
+        ),
         :pageviews
       )
     end
@@ -318,11 +334,11 @@ defmodule Plausible.Exports do
 
   defp export_visitors_q(site_id, timezone, date_range) do
     visitors_sessions_q =
-      from s in sampled("sessions_v2", date_range),
-        where: s.site_id == ^site_id,
+      from s in sampled("sessions_v2"),
+        where: ^export_filter(site_id, date_range),
         group_by: selected_as(:date),
         select: %{
-          date: date(s.start, ^timezone),
+          date: date(s.timestamp, ^timezone),
           bounces: bounces(s),
           visits: visits(s),
           visit_duration: visit_duration(s),
@@ -330,8 +346,8 @@ defmodule Plausible.Exports do
         }
 
     visitors_events_q =
-      from e in sampled("events_v2", date_range),
-        where: e.site_id == ^site_id,
+      from e in sampled("events_v2"),
+        where: ^export_filter(site_id, date_range),
         group_by: selected_as(:date),
         select: %{
           date: date(e.timestamp, ^timezone),
@@ -362,8 +378,8 @@ defmodule Plausible.Exports do
   end
 
   defp export_sources_q(site_id, timezone, date_range) do
-    from s in sampled("sessions_v2", date_range),
-      where: s.site_id == ^site_id,
+    from s in sampled("sessions_v2"),
+      where: ^export_filter(site_id, date_range),
       group_by: [
         selected_as(:date),
         selected_as(:source),
@@ -376,7 +392,7 @@ defmodule Plausible.Exports do
       ],
       order_by: selected_as(:date),
       select: [
-        date(s.start, ^timezone),
+        date(s.timestamp, ^timezone),
         selected_as(s.referrer_source, :source),
         s.referrer,
         s.utm_source,
@@ -393,8 +409,9 @@ defmodule Plausible.Exports do
   end
 
   defp export_pages_q(site_id, timezone, date_range) do
-    from e in sampled("events_v2", date_range),
-      where: [site_id: ^site_id, name: "pageview"],
+    from e in sampled("events_v2"),
+      where: ^export_filter(site_id, date_range),
+      where: [name: "pageview"],
       group_by: [selected_as(:date), e.pathname],
       order_by: selected_as(:date),
       select: [
@@ -411,16 +428,16 @@ defmodule Plausible.Exports do
   end
 
   defp export_entry_pages_q(site_id, timezone, date_range) do
-    from s in sampled("sessions_v2", date_range),
-      where: s.site_id == ^site_id,
+    from s in sampled("sessions_v2"),
+      where: ^export_filter(site_id, date_range),
       group_by: [selected_as(:date), s.entry_page],
       order_by: selected_as(:date),
       select: [
-        date(s.start, ^timezone),
+        date(s.timestamp, ^timezone),
         s.entry_page,
         visitors(s),
         selected_as(
-          fragment("toUInt64(round(sum(?)*any(_sample_factor)))", s.sign),
+          fragment("toUInt64(round(greatest(sum(?),0)*any(_sample_factor)))", s.sign),
           :entrances
         ),
         visit_duration(s),
@@ -430,17 +447,17 @@ defmodule Plausible.Exports do
   end
 
   defp export_exit_pages_q(site_id, timezone, date_range) do
-    from s in sampled("sessions_v2", date_range),
-      where: s.site_id == ^site_id,
+    from s in sampled("sessions_v2"),
+      where: ^export_filter(site_id, date_range),
       group_by: [selected_as(:date), s.exit_page],
       order_by: selected_as(:date),
       select: [
-        date(s.start, ^timezone),
+        date(s.timestamp, ^timezone),
         s.exit_page,
         visitors(s),
         visit_duration(s),
         selected_as(
-          fragment("toUInt64(round(sum(?)*any(_sample_factor)))", s.sign),
+          fragment("toUInt64(round(greatest(sum(?),0)*any(_sample_factor)))", s.sign),
           :exits
         ),
         bounces(s),
@@ -449,13 +466,13 @@ defmodule Plausible.Exports do
   end
 
   defp export_locations_q(site_id, timezone, date_range) do
-    from s in sampled("sessions_v2", date_range),
-      where: s.site_id == ^site_id,
-      where: s.city_geoname_id != 0 and s.country_code != "\0\0" and s.country_code != "ZZ",
-      group_by: [selected_as(:date), s.country_code, selected_as(:region), s.city_geoname_id],
+    from s in sampled("sessions_v2"),
+      where: ^export_filter(site_id, date_range),
+      where: s.country_code != "\0\0" and s.country_code != "ZZ",
+      group_by: [selected_as(:date), s.country_code, s.subdivision1_code, s.city_geoname_id],
       order_by: selected_as(:date),
       select: [
-        date(s.start, ^timezone),
+        date(s.timestamp, ^timezone),
         selected_as(s.country_code, :country),
         selected_as(s.subdivision1_code, :region),
         selected_as(s.city_geoname_id, :city),
@@ -468,12 +485,12 @@ defmodule Plausible.Exports do
   end
 
   defp export_devices_q(site_id, timezone, date_range) do
-    from s in sampled("sessions_v2", date_range),
-      where: s.site_id == ^site_id,
+    from s in sampled("sessions_v2"),
+      where: ^export_filter(site_id, date_range),
       group_by: [selected_as(:date), s.screen_size],
       order_by: selected_as(:date),
       select: [
-        date(s.start, ^timezone),
+        date(s.timestamp, ^timezone),
         selected_as(s.screen_size, :device),
         visitors(s),
         visits(s),
@@ -484,12 +501,12 @@ defmodule Plausible.Exports do
   end
 
   defp export_browsers_q(site_id, timezone, date_range) do
-    from s in sampled("sessions_v2", date_range),
-      where: s.site_id == ^site_id,
+    from s in sampled("sessions_v2"),
+      where: ^export_filter(site_id, date_range),
       group_by: [selected_as(:date), s.browser, s.browser_version],
       order_by: selected_as(:date),
       select: [
-        date(s.start, ^timezone),
+        date(s.timestamp, ^timezone),
         s.browser,
         s.browser_version,
         visitors(s),
@@ -501,12 +518,12 @@ defmodule Plausible.Exports do
   end
 
   defp export_operating_systems_q(site_id, timezone, date_range) do
-    from s in sampled("sessions_v2", date_range),
-      where: s.site_id == ^site_id,
+    from s in sampled("sessions_v2"),
+      where: ^export_filter(site_id, date_range),
       group_by: [selected_as(:date), s.operating_system, s.operating_system_version],
       order_by: selected_as(:date),
       select: [
-        date(s.start, ^timezone),
+        date(s.timestamp, ^timezone),
         s.operating_system,
         s.operating_system_version,
         visitors(s),