+++ lib/plausible/stats/query.ex
@@ -14,7 +14,7 @@ defmodule Plausible.Stats.Query do
             experimental_reduced_joins?: false
 
   require OpenTelemetry.Tracer, as: Tracer
-  alias Plausible.Stats.{Filters, Interval}
+  alias Plausible.Stats.{Filters, Interval, Imported}
 
   @type t :: %__MODULE__{}
 
@@ -31,7 +31,6 @@ defmodule Plausible.Stats.Query do
       |> put_interval(params)
       |> put_parsed_filters(params)
       |> put_imported_opts(site, params)
-      |> maybe_drop_prop_filter(site)
 
     on_ee do
       query = Plausible.Stats.Sampling.put_threshold(query, params)
@@ -151,7 +150,7 @@ defmodule Plausible.Stats.Query do
 
   defp put_period(query, site, %{"period" => "all"}) do
     now = today(site.timezone)
-    start_date = Plausible.Sites.local_start_date(site) || now
+    start_date = Plausible.Sites.stats_start_date(site) || now
 
     struct!(query,
       period: "all",
@@ -269,25 +268,12 @@ defmodule Plausible.Stats.Query do
     )
   end
 
-  defp maybe_drop_prop_filter(query, site) do
-    prop_filter? = Map.has_key?(query.filters, "props")
-
-    props_available? = fn ->
-      site = Plausible.Repo.preload(site, :owner)
-      Plausible.Billing.Feature.Props.check_availability(site.owner) == :ok
-    end
-
-    if prop_filter? && !props_available?.(),
-      do: struct!(query, filters: Map.drop(query.filters, ["props"])),
-      else: query
-  end
-
   @spec include_imported?(t(), Plausible.Site.t(), boolean()) :: boolean()
   def include_imported?(query, site, requested?) do
     cond do
       is_nil(site.latest_import_end_date) -> false
       Date.after?(query.date_range.first, site.latest_import_end_date) -> false
-      Enum.any?(query.filters) -> false
+      not Imported.schema_supports_query?(query) -> false
       query.period == "realtime" -> false
       true -> requested?
     end