+++ lib/plausible_web/controllers/api/external_stats_controller.ex
@@ -6,8 +6,7 @@ defmodule PlausibleWeb.Api.ExternalStatsController do
 
   def realtime_visitors(conn, _params) do
     site = conn.assigns.site
-    query = Query.from(site, %{"period" => "realtime"})
-    json(conn, Plausible.Stats.Clickhouse.current_visitors(site, query))
+    json(conn, Plausible.Stats.current_visitors(site))
   end
 
   def aggregate(conn, params) do
@@ -193,15 +192,7 @@ defmodule PlausibleWeb.Api.ExternalStatsController do
     end
   end
 
-  defp validate_metric("events" = metric, query) do
-    if query.include_imported do
-      {:error, "Metric `#{metric}` cannot be queried with imported data"}
-    else
-      {:ok, metric}
-    end
-  end
-
-  defp validate_metric(metric, _) when metric in ["visitors", "pageviews"] do
+  defp validate_metric(metric, _) when metric in ["visitors", "pageviews", "events"] do
     {:ok, metric}
   end