+++ lib/plausible_web/controllers/api/stats_controller.ex
@@ -12,8 +12,7 @@ defmodule PlausibleWeb.Api.StatsController do
 
   @revenue_metrics on_ee(do: Plausible.Stats.Goal.Revenue.revenue_metrics(), else: [])
 
-  plug :date_validation_plug
-  plug PlausibleWeb.Plugs.InjectDebugHeaders
+  plug(:date_validation_plug)
 
   @doc """
   Returns a time-series based on given parameters.
@@ -1205,9 +1204,7 @@ defmodule PlausibleWeb.Api.StatsController do
 
     params = Map.put(params, "property", prefixed_prop)
 
-    query =
-      Query.from(site, params)
-      |> Map.put(:include_imported, false)
+    query = Query.from(site, params)
 
     metrics =
       if query.filters["event:goal"] do