+++ lib/plausible_web/live/imports_exports_settings.ex
@@ -37,13 +37,10 @@ defmodule PlausibleWeb.Live.ImportsExportsSettings do
       |> assign_new(:current_user, fn ->
         Plausible.Repo.get(Plausible.Auth.User, user_id)
       end)
-      |> assign_new(:max_imports, fn %{site: site} ->
-        Imported.max_complete_imports(site)
-      end)
 
     :ok = Imported.listen()
 
-    {:ok, socket}
+    {:ok, assign(socket, max_imports: Imported.max_complete_imports())}
   end
 
   def render(assigns) do
@@ -54,8 +51,6 @@ defmodule PlausibleWeb.Live.ImportsExportsSettings do
       )
 
     at_maximum? = length(assigns.site_imports) >= assigns.max_imports
-
-    csv_imports_exports_enabled? = FunWithFlags.enabled?(:csv_imports_exports, for: assigns.site)
 
     import_warning =
       cond do
@@ -74,8 +69,7 @@ defmodule PlausibleWeb.Live.ImportsExportsSettings do
       assign(assigns,
         import_in_progress?: import_in_progress?,
         at_maximum?: at_maximum?,
-        import_warning: import_warning,
-        csv_imports_exports_enabled?: csv_imports_exports_enabled?
+        import_warning: import_warning
       )
 
     ~H"""
@@ -90,7 +84,6 @@ defmodule PlausibleWeb.Live.ImportsExportsSettings do
       </.button_link>
 
       <.button_link
-        :if={@csv_imports_exports_enabled?}
         class="w-36 h-20"
         theme="bright"
         disabled={@import_in_progress? or @at_maximum?}
@@ -121,74 +114,87 @@ defmodule PlausibleWeb.Live.ImportsExportsSettings do
     </div>
     <ul :if={not Enum.empty?(@site_imports)}>
       <li :for={entry <- @site_imports} class="py-4 flex items-center justify-between space-x-4">
-        <div class="flex flex-col">
-          <div class="flex items-center text-sm leading-5 font-medium text-gray-900 dark:text-gray-100">
-            <Heroicons.clock
-              :if={entry.live_status == SiteImport.pending()}
-              class="block h-6 w-5 text-indigo-600 dark:text-green-600"
-            />
-            <.spinner
-              :if={entry.live_status == SiteImport.importing()}
-              class="block h-6 w-5 text-indigo-600 dark:text-green-600"
-            />
-            <Heroicons.check
-              :if={entry.live_status == SiteImport.completed()}
-              class="block h-6 w-5 text-indigo-600 dark:text-green-600"
-            />
-            <Heroicons.exclamation_triangle
-              :if={entry.live_status == SiteImport.failed()}
-              class="block h-6 w-5 text-red-700 dark:text-red-700"
-            />
-            <div :if={entry.live_status == SiteImport.failed()}>
-              Import failed -
-            </div>
-            <.tooltip :if={entry.tooltip} wrapper_class="ml-2 grow" class="justify-left">
-              <%= Plausible.Imported.SiteImport.label(entry.site_import) %>
-              <:tooltip_content>
-                <.notice_message message_label={entry.tooltip} />
-              </:tooltip_content>
-            </.tooltip>
-            <div :if={!entry.tooltip} class="ml-2">
-              <%= Plausible.Imported.SiteImport.label(entry.site_import) %>
-            </div>
-            <div :if={entry.live_status == SiteImport.completed()} class="text-xs font-normal ml-1">
-              (<%= PlausibleWeb.StatsView.large_number_format(
-                pageview_count(entry.site_import, @pageview_counts)
-              ) %> page views)
-            </div>
-          </div>
-          <div class="text-sm leading-5 text-gray-500 dark:text-gray-200">
-            From <%= format_date(entry.site_import.start_date) %> to <%= format_date(
-              entry.site_import.end_date
-            ) %>
-            <%= if entry.live_status == SiteImport.completed() do %>
-              (imported
-            <% else %>
-              (started
-            <% end %>
-            on <%= format_date(entry.site_import.inserted_at) %>)
-          </div>
-        </div>
-        <.button
-          data-to={"/#{URI.encode_www_form(@site.domain)}/settings/forget-import/#{entry.site_import.id}"}
-          theme="danger"
-          data-method="delete"
-          data-csrf={Plug.CSRFProtection.get_csrf_token()}
-          class="sm:ml-3 sm:w-auto w-full"
-          data-confirm="Are you sure you want to delete this import?"
-        >
-          <span :if={entry.live_status == SiteImport.completed()}>
-            Delete Import
-          </span>
-          <span :if={entry.live_status == SiteImport.failed()}>
-            Discard
-          </span>
-          <span :if={entry.live_status not in [SiteImport.completed(), SiteImport.failed()]}>
-            Cancel Import
-          </span>
-        </.button>
+        <.import_entry entry={entry} site={@site} pageview_counts={@pageview_counts} />
       </li>
     </ul>
+    """
+  end
+
+  defp import_entry(assigns) do
+    label_class =
+      if assigns.entry.live_status != SiteImport.failed() do
+        "ml-2"
+      end
+
+    assigns = assign(assigns, :label_class, label_class)
+
+    ~H"""
+    <div class="flex flex-col">
+      <div class="flex items-center text-sm leading-5 font-medium text-gray-900 dark:text-gray-100">
+        <Heroicons.clock
+          :if={@entry.live_status == SiteImport.pending()}
+          class="block h-6 w-5 text-indigo-600 dark:text-green-600"
+        />
+        <.spinner
+          :if={@entry.live_status == SiteImport.importing()}
+          class="block h-6 w-5 text-indigo-600 dark:text-green-600"
+        />
+        <Heroicons.check
+          :if={@entry.live_status == SiteImport.completed()}
+          class="block h-6 w-5 text-indigo-600 dark:text-green-600"
+        />
+        <Heroicons.exclamation_triangle
+          :if={@entry.live_status == SiteImport.failed()}
+          class="block h-6 w-5 text-red-700 dark:text-red-700"
+        />
+        <div :if={@entry.live_status == SiteImport.failed()} class="ml-2 mr-1">
+          Import failed -
+        </div>
+        <.tooltip :if={@entry.tooltip} wrapper_class={[@label_class, "grow"]} class="justify-left">
+          <%= Plausible.Imported.SiteImport.label(@entry.site_import) %>
+          <:tooltip_content>
+            <.notice_message message_label={@entry.tooltip} />
+          </:tooltip_content>
+        </.tooltip>
+        <div :if={!@entry.tooltip} class={[@label_class]}>
+          <%= Plausible.Imported.SiteImport.label(@entry.site_import) %>
+        </div>
+        <div :if={@entry.live_status == SiteImport.completed()} class="text-xs font-normal ml-1">
+          (<%= PlausibleWeb.StatsView.large_number_format(
+            pageview_count(@entry.site_import, @pageview_counts)
+          ) %> page views)
+        </div>
+      </div>
+      <div class="text-sm leading-5 text-gray-500 dark:text-gray-200">
+        From <%= format_date(@entry.site_import.start_date) %> to <%= format_date(
+          @entry.site_import.end_date
+        ) %>
+        <%= if @entry.live_status == SiteImport.completed() do %>
+          (imported
+        <% else %>
+          (started
+        <% end %>
+        on <%= format_date(@entry.site_import.inserted_at) %>)
+      </div>
+    </div>
+    <.button
+      data-to={"/#{URI.encode_www_form(@site.domain)}/settings/forget-import/#{@entry.site_import.id}"}
+      theme="danger"
+      data-method="delete"
+      data-csrf={Plug.CSRFProtection.get_csrf_token()}
+      class="sm:ml-3 sm:w-auto w-full"
+      data-confirm="Are you sure you want to delete this import?"
+    >
+      <span :if={@entry.live_status == SiteImport.completed()}>
+        Delete Import
+      </span>
+      <span :if={@entry.live_status == SiteImport.failed()}>
+        Discard
+      </span>
+      <span :if={@entry.live_status not in [SiteImport.completed(), SiteImport.failed()]}>
+        Cancel Import
+      </span>
+    </.button>
     """
   end